+++ examples/hooks-app/config/pontTemplate.ts
@@ -1,13 +1,13 @@
-import { Interface, BaseClass, Property, CodeGenerator } from 'pont-engine';
+import { Interface, BaseClass, Property, CodeGenerator } from "pont-engine";
 
 export default class MyGenerator extends CodeGenerator {
   getInterfaceContentInDeclaration(inter: Interface) {
     const method = inter.method.toUpperCase();
 
     const paramsCode = inter
-      .getParamsCode('Params')
-      .replace('lock: number', 'lock?: number')
-      .replace(': file', ': FormData');
+      .getParamsCode("Params")
+      .replace("lock: number", "lock?: number")
+      .replace(": file", ": FormData");
 
     return `
       export ${paramsCode}
@@ -21,7 +21,7 @@ export default class MyGenerator extends CodeGenerator {
       export function trigger(params?: HooksParams, shouldRevalidate = true);
 
       ${
-        method === 'GET'
+        method === "GET"
           ? `
         export function useRequest(params?: HooksParams, options?: ConfigInterface): { isLoading: boolean; data: Response, error: Error };`
           : `
@@ -72,20 +72,24 @@ export default class MyGenerator extends CodeGenerator {
 
     import { PontCore } from '../../pontCore'
 
-    export ${inter.getParamsCode('Params', this.surrounding)}
+    export ${inter.getParamsCode("Params", this.surrounding)}
 
     export const method = "${method}";
 
     export function mutate(params = {}, newValue = undefined, shouldRevalidate = true) {
-      return SWR.mutate(Hooks.getUrlKey("${inter.path}", params, "${method}"), newValue, shouldRevalidate);
+      return SWR.mutate(Hooks.getUrlKey("${
+        inter.path
+      }", params, "${method}"), newValue, shouldRevalidate);
     }
 
     export function trigger(params = {}, shouldRevalidate = true) {
-      return SWR.trigger(Hooks.getUrlKey("${inter.path}", params, "${method}"), shouldRevalidate);
+      return SWR.trigger(Hooks.getUrlKey("${
+        inter.path
+      }", params, "${method}"), shouldRevalidate);
     }
 
     ${
-      method === 'GET'
+      method === "GET"
         ? `
       export function useRequest(params = {}, swrOptions = {}) {
         return Hooks.useRequest("${inter.path}", params, swrOptions);
@@ -98,7 +102,9 @@ export default class MyGenerator extends CodeGenerator {
     }
 
     export function request(params = {}, option  = {}) {
-      return PontCore.fetch(PontCore.getUrl("${inter.path}", params, "${method}"), {
+      return PontCore.fetch(PontCore.getUrl("${
+        inter.path
+      }", params, "${method}"), {
         ...option,
         method: "${method}",
       });